requirements.txt
folium
geocoder
django
django-cors-headers
django-crispy-forms
django_plotly_dash
djangorestframework
django-admin-tools
django-translation-manager
django-modeltranslation
django-rosetta
channels_redis
dash_renderer